streamlit/requirements.txt
streamlit
pandas
databricks-sql-connector
python-dotenv